testing/web-platform/mozilla/tests/focus/focus-before-iframe-loaded-different-site.html

<!doctype html>
<meta charset=utf-8>
<title>focus() before iframe loaded different site</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
setup({explicit_done:true});
window.onmessage = function(e) {
    test(function() {
        assert_equals(e.data, "PASS", "Check verdict");
    }, "Check result");
    w.close();
    done();
};
var w = window.open("support/focus-before-iframe-loaded-different-site-outer.sub.html");
</script>
